Created common bundle for DB wire components
