+++ kura/distrib/config/kura.build.properties
@@ -87,6 +87,7 @@ org.eclipse.kura.web2.version=2.2.0-SNAPSHOT
 org.eclipse.kura.wire.camel.version=1.2.0-SNAPSHOT
 org.eclipse.kura.wire.component.provider.version=1.2.0-SNAPSHOT
 org.eclipse.kura.wire.h2db.component.provider.version=2.2.0-SNAPSHOT
+org.eclipse.kura.wire.basedb.component.provider.version=1.0.0-SNAPSHOT
 org.eclipse.kura.wire.script.filter.provider.version=1.2.0-SNAPSHOT
 org.eclipse.kura.wire.helper.provider.version=1.2.0-SNAPSHOT
 org.eclipse.kura.wire.provider.version=1.2.0-SNAPSHOT
