+++ kura/org.eclipse.kura.wire.basedb.component.provider/META-INF/MANIFEST.MF
@@ -1,8 +1,8 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: H2 Database Wire Components Provider
-Bundle-SymbolicName: org.eclipse.kura.wire.h2db.component.provider;singleton:=true
-Bundle-Version: 2.2.0.qualifier
+Bundle-Name: Base Database Wire Components Provider
+Bundle-SymbolicName: org.eclipse.kura.wire.basedb.component.provider;singleton:=true
+Bundle-Version: 1.0.0.qualifier
 Bundle-Vendor: Eclipse Kura
 Bundle-License: Eclipse Public License v2.0
 Bundle-Category: Kura Wires
@@ -14,7 +14,6 @@ Import-Package: org.apache.logging.log4j;version="2.8.2",
  org.eclipse.kura.type;version="[1.0,2.0)",
  org.eclipse.kura.util.collection;version="[1.0,2.0)",
  org.eclipse.kura.wire;version="[2.0,3.0)",
- org.eclipse.kura.wire.basedb.store;version="[2.0,3.0)",
  org.osgi.framework;version="1.8.0",
  org.osgi.service.component;version="1.2.0",
  org.osgi.service.wireadmin;version="1.0.1",
@@ -22,5 +21,5 @@ Import-Package: org.apache.logging.log4j;version="2.8.2",
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Service-Component: OSGI-INF/*.xml
-Export-Package: org.eclipse.kura.internal.wire.h2db.filter,
- org.eclipse.kura.internal.wire.h2db.store;version="2.0.0"
+Export-Package: org.eclipse.kura.wire.basedb.filter,
+ org.eclipse.kura.wire.basedb.store;version="2.0.0"
