+++ kura/org.eclipse.kura.wire.basedb.component.provider/build.properties
@@ -1,0 +1,8 @@
+source.. = src/main/java/,\
+           src/main/resources/,\
+           src/test/java/,\
+           src/test/resources/
+bin.includes = META-INF/,\
+               .,\
+               about.html
+additional.bundles = org.apache.logging.log4j.api
