+++ kura/org.eclipse.kura.wire.h2db.component.provider/src/main/java/org/eclipse/kura/internal/wire/h2db/store/H2DbWireRecordStoreOptions.java
@@ -13,74 +13,17 @@
  *******************************************************************************/
 package org.eclipse.kura.internal.wire.h2db.store;
 
-import static java.util.Objects.nonNull;
-import static java.util.Objects.requireNonNull;
-
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
+
+import org.eclipse.kura.wire.basedb.store.BaseDbWireRecordStoreOptions;
 
 /**
  * The Class DbWireRecordStoreOptions is responsible to contain all the DB Wire
  * Record Store related options
  */
-final class H2DbWireRecordStoreOptions {
+final class H2DbWireRecordStoreOptions extends BaseDbWireRecordStoreOptions {
 
-    private static final int DEFAULT_MAXIMUM_TABLE_SIZE = 10000;
-
-    protected static final String MAXIMUM_TABLE_SIZE = "maximum.table.size";
-
-    protected static final String CLEANUP_RECORDS_KEEP = "cleanup.records.keep";
-
-    protected static final String TABLE_NAME = "table.name";
-
-    private final Map<String, Object> properties;
-
-    /**
-     * Instantiates a new DB wire record store options.
-     *
-     * @param properties
-     *            the configured properties
-     */
-    H2DbWireRecordStoreOptions(final Map<String, Object> properties) {
-        requireNonNull(properties, "Properties cannot be null");
-        this.properties = Collections.unmodifiableMap(new HashMap<>(properties));
-    }
-
-    /**
-     * Returns the number of records to keep as configured .
-     *
-     * @return the number of records
-     */
-    int getNoOfRecordsToKeep() {
-        int noOfRecords = 5000;
-        final Object cleanUp = this.properties.get(CLEANUP_RECORDS_KEEP);
-        if (nonNull(cleanUp) && cleanUp instanceof Integer) {
-            noOfRecords = (Integer) cleanUp;
-        }
-        return noOfRecords;
-    }
-
-    int getMaximumTableSize() {
-        int maximumSize = DEFAULT_MAXIMUM_TABLE_SIZE;
-        final Object propertiesMaximumSize = this.properties.get(MAXIMUM_TABLE_SIZE);
-        if (nonNull(propertiesMaximumSize) && propertiesMaximumSize instanceof Integer) {
-            maximumSize = (Integer) propertiesMaximumSize;
-        }
-        return maximumSize;
-    }
-
-    /**
-     * Returns the name of the table as configured.
-     *
-     * @return the name of the table
-     */
-    String getTableName() {
-        String tableName = null;
-        final Object name = this.properties.get(TABLE_NAME);
-        if (nonNull(name) && name instanceof String) {
-            tableName = name.toString();
-        }
-        return tableName;
+    protected H2DbWireRecordStoreOptions(final Map<String, Object> properties) {
+        super(properties);
     }
 }
