+++ kura/pom.xml
@@ -87,6 +87,7 @@
         <module>org.eclipse.kura.wire.camel</module>
         <module>org.eclipse.kura.wire.script.filter.provider</module>
         <module>org.eclipse.kura.wire.h2db.component.provider</module>
+        <module>org.eclipse.kura.wire.basedb.component.provider</module>
         <module>org.eclipse.kura.wire.component.provider</module>
         <module>org.eclipse.kura.wire.helper.provider</module>
         <module>org.eclipse.kura.wire.provider</module>
