+++ kura/test/org.eclipse.kura.wire.h2db.component.provider.test/META-INF/MANIFEST.MF
@@ -17,4 +17,4 @@ Import-Package: org.eclipse.kura.core.testutil;version="1.0.0",
  org.mockito.invocation;version="1.10.19",
  org.mockito.stubbing;version="1.10.19",
  org.slf4j;version="1.6.4"
-Fragment-Host: org.eclipse.kura.wire.h2db.component.provider;bundle-version="2.1.0"
+Fragment-Host: org.eclipse.kura.wire.h2db.component.provider;bundle-version="2.2.0"
